% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCount.R
\name{exportCount}
\alias{exportCount}
\title{write computeMatrix output}
\usage{
exportCount(se, filename = tempfile())
}
\arguments{
\item{se}{a SummarizedExperiment object from \link{importCount}.}

\item{filename}{output filename}
}
\value{
zipped output filename
}
\description{
Write a SummarizedExperiment object into computeMatrix output.
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
se <- importCount(file)
exportCount(se)
}
