% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCount.R
\name{filterCount}
\alias{filterCount}
\title{filter computeMatrix output}
\usage{
filterCount(se, subset)
}
\arguments{
\item{se}{a SummarizedExperiment object from \link{importCount}.}

\item{subset}{filter condition for rows.}
}
\value{
a SummarizedExperiment object
}
\description{
Filter output of importCount.
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
se <- importCount(file)
library(SummarizedExperiment)
keep <- rowMeans(assays(se)[[1]], na.rm = TRUE) < 2 ## arbitory number
nrow(se)
se <- filterCount(se, subset=keep)
nrow(se)
table(keep)
}
