% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCount.R
\name{importCount}
\alias{importCount}
\title{read computeMatrix output}
\usage{
importCount(file)
}
\arguments{
\item{file}{count filename of computeMatrix output.}
}
\value{
a SummarizedExperiment object
}
\description{
Read computeMatrix output into a SummarizedExperiment object
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
se <- importCount(file)
}
