% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCounts.R
\name{mergeCounts}
\alias{mergeCounts}
\title{merge profile from multiple counts file}
\usage{
mergeCounts(files, FUN = colMeans, na.rm = TRUE)
}
\arguments{
\item{files}{count filenames of computeMatrix output.}

\item{FUN}{function for each column in counts. Could set to 
colMeans, colSum.}

\item{na.rm}{logical. Should missing values (including NaN) be 
omitted from the calculations?}
}
\value{
data.frame object.
}
\description{
Merge profile from multiple counts file by given function
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
mergeCounts(file)
}
