% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plot heatmap}
\usage{
plotHeatmap(
  se,
  facet = as.formula("group~sample"),
  transformFUN = log2,
  pseudo = 1,
  orderFUN = rowMeans,
  orderBy = 1,
  sampleOrder = NULL,
  groupOrder = NULL,
  fill_gradient = scale_fill_gradient(low = "blue", high = "red"),
  boderColor = NA,
  xaxis_breaks = NULL,
  xaxis_label = NULL,
  yaxis_breaks = NULL,
  yaxis_label = NULL
)
}
\arguments{
\item{se}{a SummarizedExperiment object from \link{importCount}.}

\item{facet}{This will be passed to 
\link[ggplot2:facet_grid]{facet_grid}:rows.
It can be set to NULL.}

\item{transformFUN}{The transform function of values.}

\item{pseudo}{Pseudo value will be add to values before apply transformFUN.}

\item{orderFUN}{function used to reorder Y axis of the heatmap}

\item{orderBy}{the sample name or index used to order Y axis}

\item{sampleOrder, groupOrder}{order of the samples(column) or groups(row).}

\item{fill_gradient}{color fill gradient function. 
see \link[ggplot2:scale_colour_gradient]{scale_colour_gradient}.}

\item{boderColor}{The color of the heatmap cell border.}

\item{xaxis_breaks, xaxis_label}{xaxis breaks and labels.
see \link[ggplot2:scale_continuous]{scale_x_continuous}.}

\item{yaxis_breaks, yaxis_label}{xaxis breaks and labels.
see \link[ggplot2:scale_x_discrete]{scale_y_discrete}.}
}
\value{
ggplot object
}
\description{
Plot a heatmap for counts.
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
se <- importCount(file)
plotHeatmap(se, yaxis_breaks="100033817", yaxis_label="geneA")
}
