% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plotProfile}
\alias{plotProfile}
\title{plot profiles}
\usage{
plotProfile(
  se,
  loessSmooth = TRUE,
  span = 1/25,
  facet = "group",
  xaxis_breaks = NULL,
  xaxis_label = NULL
)
}
\arguments{
\item{se}{a SummarizedExperiment object from \link{importCount}.}

\item{loessSmooth}{Use \link[stats:scatter.smooth]{loess.smooth} 
to smooth curve or not.}

\item{span}{smoothness parameter for loess.}

\item{facet}{group or sample. This will be passed to \link[ggplot2:facet_wrap]{facet_wrap}.}

\item{xaxis_breaks, xaxis_label}{xaxis breaks and labels. see \link[ggplot2:scale_continuous]{scale_x_continuous}.}
}
\value{
ggplot object
}
\description{
Plot a profile for counts.
}
\examples{
file <- system.file("extdata", "count.gz", package= "deepToolsDownstream")
se <- importCount(file)
plotProfile(se)
}
