% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabels.R
\name{relabels}
\alias{relabels}
\title{change sample/group labels}
\usage{
relabels(se, attr = "sample_labels", values)
}
\arguments{
\item{se}{a SummarizedExperiment object from \link{importCount}.}

\item{attr}{attribute to be changed. Could be sample_labels or group_labels.}

\item{values}{values to be set.}
}
\value{
a SummarizedExperiment object
}
\description{
Change sample/group labels of imported data.
}
